# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Rest Clon de Airbnb",
        "version": "1.0.0",
        "contact": {
            "email": "blu.r@outlook.com"
        }
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "summary": "Obtener todos los usuarios",
                "description": " ",
                "tags": ["Users"],
                "responses": {
                    "200": {
                        "description": "Obtenemos exitosamente todos los usuarios"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la petición no se resuelva"
                    }
                }
            }
        },
...
        "/api/v1/accommodations": {
            "get": {
                "summary": "Obtener todos los alojamientos",
                "description": " ",
                "tags": ["Accommodations"],
                "responses": {
                    "200": {
                        "description": "Obtenemos exitosamente todos los alojamientos"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la petición no se resuelva"
                    }
                }
            },
            "post": {
                "summary": "Crear un alojamiento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "jwt",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "UUID"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Objeto con información del alojamiento",
                        "required": true,
                        "type": "object"
                    }
                ],
                "description": "Solo puede realizarlo un usuario 'host'",
                "tags": ["Accommodations"],
                "responses": {
                    "201": {
                        "description": "Creamos exitosamente un alojamiento"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la petición no se resuelva"
                    }
                }
            }
        },
...
        "/api/v1/reservations": {
            "get": {
                "summary": "Obtener todas las reservaciones",
                "description": " ",
                "tags": ["Reservations"],
                "responses": {
                    "200": {
                        "description": "Obtenemos exitosamente todas las reservaciones"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la petición no se resuelva"
                    }
                }
            }
        },
...
    }
}